refactor(scripts): extract cleanup CLI boilerplate to kill CPD duplication
The SonarCloud gate flagged 5.0% duplicated new-code lines: the main()
opening (--apply mode parse + kennel lookup + targeting log) in the new
cleanup-mh3-mn-test-events-v2.ts matched the same block in the pre-existing
v1 script. Extract it to scripts/lib/cleanup-cli.ts (parseApplyMode +
resolveCleanupKennel, split into two small functions so the shared helper
doesn't itself re-trip CPD), and refactor both new cleanup scripts to use it.
